requirements.txt
flask==1.1.2
flask-admin==1.5.7
flask-migrate==2.7.0
python-dotenv==0.15.0
pillow==8.1.2
flask-admin-s3-upload==0.1.4